Consolidate GitHub integration into self-contained dojo/github/ package
Collapse dojo/github.py and dojo/github_issue_link/ into a single
dojo/github/ package matching the canonical dojo/url/ reference layout
described in CLAUDE.md (models, admin, services, ui/forms, ui/views,
ui/urls, templates). Backward-compat re-exports left in dojo/models.py
and dojo/forms.py. No behavior change, no migrations.
