+++ dojo/github/__init__.py
@@ -1,0 +1,1 @@
+import dojo.github.admin  # noqa: F401
