+++ dojo/github/templates/dojo/delete_github.html
